third_party/WebKit/LayoutTests/device_orientation/orientation/create-event.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
description('Tests that document.createEvent() works with DeviceOrientationEvent.');

var event = document.createEvent('DeviceOrientationEvent');
var newEvent = new CustomEvent("deviceorientation", {
    bubbles: false, cancelable: false,
    alpha: 1.0,
    beta: 2.0,
    gama: 3.0,
    absolute: false
});

shouldBeTrue("typeof event == 'object'");
shouldBe("event.__proto__", "DeviceOrientationEvent.prototype");

shouldBeTrue("'type' in event");
shouldBeTrue("'bubbles' in event");
shouldBeTrue("'cancelable' in event");
shouldBeTrue("'alpha' in event");
shouldBeTrue("'beta' in event");
shouldBeTrue("'gamma' in event");
shouldBeTrue("'absolute' in event");

shouldBeTrue("typeof event.type == 'string'");
shouldBeTrue("typeof event.bubbles == 'boolean'");
shouldBeTrue("typeof event.cancelable == 'boolean'");
shouldBeTrue("typeof event.alpha == 'object'");
shouldBeTrue("typeof event.beta == 'object'");
shouldBeTrue("typeof event.gamma == 'object'");
shouldBeTrue("typeof event.absolute == 'boolean'");

shouldBeEqualToString("newEvent.type", "deviceorientation");
shouldBeFalse("newEvent.bubbles");
shouldBeFalse("newEvent.cancelable");

</script>
</body>
</html>
