third_party/WebKit/LayoutTests/device_orientation/orientation/create-event-expected.txt

Tests that document.createEvent() works with DeviceOrientationEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof event == 'object' is true
PASS event.__proto__ is DeviceOrientationEvent.prototype
PASS 'type' in event is true
PASS 'bubbles' in event is true
PASS 'cancelable' in event is true
PASS 'alpha' in event is true
PASS 'beta' in event is true
PASS 'gamma' in event is true
PASS 'absolute' in event is true
PASS typeof event.type == 'string' is true
PASS typeof event.bubbles == 'boolean' is true
PASS typeof event.cancelable == 'boolean' is true
PASS typeof event.alpha == 'object' is true
PASS typeof event.beta == 'object' is true
PASS typeof event.gamma == 'object' is true
PASS typeof event.absolute == 'boolean' is true
PASS newEvent.type is "deviceorientation"
PASS newEvent.bubbles is false
PASS newEvent.cancelable is false
PASS successfullyParsed is true

TEST COMPLETE
